[OR1K] Translate non-RMW atomics to non-atomic instructions.
l.lwa/l.swa are only appropriate in an LL/SC loop, in particular
l.swa without a matching l.lwa is not guaranteed to store anything.

For word-size atomic loads or stores, a regular load or store
followed and/or preceded by a barrier is appropriate, but it does not
appear that OR1K defines any barriers appropriate in a multiprocessor
system, and they are not necessary on uniprocessor systems anyhow,
so for now do not translate any.